Tidy up QDateTimeParser in assorted petty ways
A bit-field was necessarily equal to either zero or its (single-bit)
mask; so comparing to the mask was superfluous.

A local value set and then conditionally over-ridden could easily be
made const by using the condition in a ?: assignment.

A use of qMin() lent itself to turning into a simple conditional
assignment.

Change-Id: I541d98044b23b8c51e6169f617a5841ba8fffc55
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>