QDateTimeParser: rename some confusingly-named variables
The parse() method's so-called "currentValue" is a caller-supplied
const that isn't (because it can't be) updated to reflect what's
parsed, albeit parse() passes it to other methods that presume it is
(I'll fix this in a later commit).  Renamed to reflect that it
actually just supplies defaults for any fields not parsed from the
string.  Then newCurrentValue is in fact the date-time we finally end
up having parsed, so rename it to reflect that.

Change-Id: Ic6add2673b5b245606dbb6415d84c18fc085e456
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>